+++ tests/test_tox_wheel.py
@@ -30,13 +30,34 @@ def test_enabled(testdir):
     assert result.ret == 0
 
 
-def test_enabled_clean(testdir):
-    result = testdir.run('tox', '-vv', '-e', 'py27', '--wheel', '--wheel-clean-build')
+def test_enabled_toxini_noclean(testdir):
+    testdir.tmpdir.join('tox.ini').write("""
+[testenv]
+wheel = true
+wheel_clean_build = false
+""")
+    result = testdir.run('tox', '-e', 'py27')
     result.stdout.fnmatch_lines([
         'GLOB wheel-make: *',
+    ])
+    assert 'cleaning up build directory ...' not in result.stdout.str()
+    assert 'cleaning up build directory ...' not in result.stderr.str()
+    assert result.ret == 0
+
+
+def test_enabled_toxini(testdir):
+    testdir.tmpdir.join('tox.ini').write("""
+[testenv]
+wheel = true
+""")
+    result = testdir.run('tox', '-vv', '-e', 'py27')
+    result.stdout.fnmatch_lines([
+        'GLOB wheel-make: *',
+        'GLOB wheel-make: cleaning up build directory ...',
         '  removing *[\\/]build',
         'GLOB finish: packaging *',
         'copying new sdistfile to *.whl*',
     ])
     assert 'is not a supported wheel on this platform.' not in result.stdout.str()
     assert 'is not a supported wheel on this platform.' not in result.stderr.str()
+