+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog
 =========
 
+0.2.0 (2019-01-11)
+------------------
+
+* Remove ``--wheel-clean-build`` CLI option. Build directory cleaning is now on by default.
+  Correct behavior should be the default.
+* Added support for ``[testenv] wheel`` (default: ``false``) and ``[testenv] wheel_clean_build`` (default: ``true``)
+  in ``tox.ini``.
+
 0.1.0 (2019-01-09)
 ------------------